backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-swagger-ui==4.11.1
psycopg[binary]>=3.2.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
SQLAlchemy>=2.0.36

